fix: 🐛 last page and home page rendering issues
If user did not have a localstorage for last page visited with an empty
string, the user did not satisfy conditions for render. If homepage was
passed into defaultHomePage function without a value it would error and
not push to the router a path. Changed from promise to simple function
to stop this incorrect functionality