fix: 🐛 get company from subsidary generating correctly
correctly extracted company_id and iso_code to make it work, and then
updated the rendering of the button to include loading states, and list
level state to prevent multiple from being clicked at the same time.
Also refactored the rendering to remove the print settings, thought it
had been done already but clearly not! Updated the location of the PEP
flag in the shareholder section